An instruction is eight bytes, and the whole of p5.js compiles
Three operands that were bytes are 16 bits, and the wide field they combine
into is 32. That is the entire change, and it takes p5.js from 1,828 compiled
functions to 4,682 - all of them. The ratchet goes 3 to 5: the bundle compiles,
fits the bytecode, and RUNS.

The caps were never a limit anyone chose. An instruction was four bytes because
that is what fits an opcode and three byte operands, and every index that did
not fit was truncated in silence: the 257th distinct property name in a
function read a different property, a function wanting more than 256 registers
aliased its own locals, and a jump past 32,767 instructions branched somewhere
that was never a target. p5.js has 1,452 declarations in its top-level frame
alone. Eight bytes costs about 8 MB on this bundle.

THE THREE CHECKS STAY, at the new widths. They are the reason this was safe to
do at all - the caps were made loud four commits ago precisely so there would
be a way to tell whether widening them had worked, and page_scripts asserted
each one as a refusal. All three now compile, and those assertions moved to
must_compile_source rather than being deleted. An operand that silently
truncates is the worst thing a compiler can have at ANY width.

One bug, and it is the one this shape of change exists to produce. patch_jump
wrote the displacement as an 8/8 split - correct for byte operands - into what
were now 16-bit halves. Every jump in the program went somewhere else. The
symptom was not a crash: it was `if` bodies silently not running, so a probe
that tested a condition and threw on failure reported success by never reaching
the throw. -Wconversion found the other forty sites; it could not find this one,
because the shift was already narrowing on purpose.

What stops p5 now is a call, at run time, which is a different kind of question
from anything so far - the first one that needs the standard library to be right
rather than the compiler.

+++ include/ctbrowser/script/vm.hpp
@@ -262,11 +262,11 @@ private:
         const function_proto * proto = nullptr;
         std::size_t ip = 0;
         std::size_t base = 0; // index into registers_ of this frame's r0
-        std::uint8_t result_reg = 0;
+        std::uint16_t result_reg = 0;
         // How many arguments ACTUALLY arrived, which is not param_count: a rest
         // parameter binds the ones past the declared list, and nothing else in
         // the frame records that they were passed.
-        std::uint8_t argc = 0;
+        std::uint16_t argc = 0;
         closure_object * closure = nullptr; // whose upvalues this body sees
         // The receiver. A JS body reads it through `this`; before this existed
         // `this` compiled to undefined unconditionally, so no method could see
@@ -299,7 +299,7 @@ private:
         std::size_t frame = 0;   // index into frames_
         std::size_t address = 0; // the catch block
         std::size_t reg_top = 0; // registers_ size on entry
-        std::uint8_t slot = 0;   // where to put the thrown value
+        std::uint16_t slot = 0;  // where to put the thrown value
     };
 
     // Run the `__fields` initialiser of `constructor` and of every class it
@@ -364,7 +364,7 @@ private:
     // so the running frame's proto belongs to a DIFFERENT program from the one
     // being executed. Subtracting its address from the wrong program's
     // functions vector gave a garbage index and read off the end of the cache.
-    flat_map<const function_proto *, flat_map<std::uint16_t, value>> string_cache_;
+    flat_map<const function_proto *, flat_map<std::uint32_t, value>> string_cache_;
     // Live try blocks, innermost last. Not per-frame, because a throw has to be
     // able to find a handler several frames up.
     std::vector<handler> handlers_;
